+++ app/decorators/reminder_decorator/base.rb
@@ -1,7 +1,8 @@
 module ReminderDecorator
   class Base < BaseDecorator
     delegate :id, :name, :valid_for_n_days, :deadline_text, :notification_text,
-             :persisted?, :slack_channel, :supervisor_slack_channel, :notify_projects_channels
+             :persisted?, :slack_channel, :supervisor_slack_channel, :notify_projects_channels,
+             :jira_issue_lead
     decorates :reminder
 
     def remind_after_days