+++ config/schedule.rb
@@ -4,3 +4,8 @@
 every "30 9 * * 1-5" do
   rake "reminders:check_all"
 end
+
+# Monday-Friday at 9:45AM
+every "45 9 * * 1-5" do
+  rake "reminders:create_jira_issues"
+end