+++ db/migrate/20170310091438_add_jira_fields_to_project_checks.rb
@@ -1,0 +1,6 @@
+class AddJiraFieldsToProjectChecks < ActiveRecord::Migration[5.0]
+  def change
+    add_column :project_checks, :jira_issue_created_at, :datetime
+    add_column :project_checks, :jira_issue_key, :string
+  end
+end