+++ spec/support/shared_examples/no_project_checks.rb
@@ -1,0 +1,5 @@
+shared_examples "no project checks" do
+  it "returns no project checks" do
+    expect(subject).to eq([])
+  end
+end