+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+# 2017-03-16
+## Added
+* Jira issues will be created 7 days before project check deadline
+* Number of days is editable for every reminder
+
 # 2017-03-09
 
 ## Added